assignment about Administrative commands
Create a Schema called "ManagementSchema."
Create a view owned by the schema that shows the annual ridership.
Create a view owned by the schema that shows the employee information
including their position and pay rate.
Create a role ManagementRole.
Give the ManagementRole select permissions on the ManagementSchema and
Exec permissions on the new employee stored procedure we created
earlier.
Create a new login for one of the employees who holds the manager
position.
Create a new user for that login.
Add that user to the Role.
Login to the database as the new User, (Remember that SQL server
authentication must be enabled for this to work.)
Run the query to view the annual ridership. Does it work?
Try to select from the table Employees. Can you do it?
Back up the database MetroAlt.